feat(ChatAI): Move ChatPanel to reusable components
