+++ package.json
@@ -5,10 +5,10 @@
     "author": "Bluesquare",
     "license": "Apache-2.0",
     "scripts": {
-        "show-lint-problems": "eslint \"./hat/assets/js/**/*.{js,jsx,ts,tsx}\" \"./plugins/polio/js/**/*.{js,jsx,ts,tsx}\"",
-        "fix-lint": "eslint \"./hat/assets/js/**/*.{js,jsx,ts,tsx}\" \"./plugins/polio/js/**/*.{js,jsx,ts,tsx}\" --cache --fix",
+        "show-lint-problems": "eslint \"./plugins/snt_malaria/js/src/**/*.{js,jsx,ts,tsx}\"",
+        "fix-lint": "eslint \"./plugins/snt_malaria/js/**/*.{js,jsx,ts,tsx}\" --cache --fix",
         "fix-lint-pre-commit": "eslint --cache --fix",
-        "export-lint-problems": "eslint \"./hat/assets/js/**/*.{js,jsx,ts,tsx}\" \"./plugins/polio/js/**/*.{js,jsx,ts,tsx}\" -f html -o ./hat/assets/js/lintErrors.html",
+        "export-lint-problems": "eslint \"./hat/assets/js/src/**/*.{js,jsx,ts,tsx}\" \"./plugins/polio/js/**/*.{js,jsx,ts,tsx}\" -f html -o ./hat/assets/js/lintErrors.html",
         "orval": "orval",
         "test": "vitest --config vitest.config.ts --project integration --project unit --project a11y --run",
         "test:api-e2e": "vitest --config vitest.config.ts --project api-e2e --run",
@@ -81,6 +81,7 @@
         "classnames": "^2.2.6",
         "color": "^3.1.2",
         "dom-to-pdf": "^0.3.1",
+        "expr-eval": "^2.0.2",
         "flume": "^1.2.0",
         "formik": "^2.2.9",
         "globals": "^16.1.0",
